requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.4
Pillow==11.1.0
